types: Export lock types
